feat:divide bottomSheetDialog into Rt & share
