+++ src/mobile/app/src/main/java/com/smilegate/Easel/presentation/view/timeline/RetweetBottomSheetDialog.kt
@@ -1,0 +1,36 @@
+package com.smilegate.Easel.presentation.view.timeline
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Button
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import com.smilegate.Easel.R
+
+class RetweetBottomSheetDialog : BottomSheetDialogFragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        super.onCreateView(inflater, container, savedInstanceState)
+        return inflater.inflate(R.layout.retweet_bottom_sheet_dialog, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val cancelButton = view.findViewById<Button>(R.id.btn_bottom_sheet_cancel)
+
+        cancelButton.setOnClickListener {
+            dismiss()
+        }
+
+    }
+
+    companion object {
+        const val TAG = "BasicBottomModalSheet"
+    }
+}
