+++ src/mobile/app/src/main/res/values/strings.xml
@@ -60,11 +60,13 @@
     <string name="ask_name_fragment_hint">사용자 아이디</string>
     <string name="ask_name_fragment_accept">가입</string>
 
-    <!-- TimelineBottomSheetDialog -->
+    <!-- RetweetBottomSheetDialog -->
     <string name="retweet_fragment_bottom_sheet">재게시</string>
     <string name="repost_fragment_bottom_sheet">인용</string>
     <string name="cancel_fragment_bottom_sheet">취소</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+
+    <!-- ShareBottomSheetDialog -->
+    <string name="link_Share_fragment_bottom_sheet">링크 복사</string>
+    <string name="Share_fragment_bottom_sheet">공유하기</string>
 
 </resources>
